# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legazpi Souvenir Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: string
        - name: isBestSeller
          in: query
          schema:
            type: boolean
        - name: isSeasonal
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"

    post:
      operationId: addToWishlist
      tags: [wishlist]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToWishlistRequest"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Product:
      type: object
      properties:
        id:
          type: string
        vendorId:
          type: string
        categoryId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        unit:
          type: string
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        isBestSeller:
          type: boolean
        isSeasonal:
          type: boolean
        isActive:
          type: boolean
        sku:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        ingredients:
          type: string
          nullable: true
        expirationMonths:
          type: integer
          nullable: true
        weight:
          type: string
          nullable: true
        vendorName:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, vendorId, categoryId, name, description, price, unit, stock, images, isBestSeller, isSeasonal, isActive, tags, rating, reviewCount, createdAt]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required: [products, total, offset, limit]
    CreateProductRequest:
      type: object
      properties:
        vendorId:
          type: string
        categoryId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        unit:
          type: string
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        isBestSeller:
          type: boolean
        isSeasonal:
          type: boolean
        sku:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        ingredients:
          type: string
          nullable: true
        expirationMonths:
          type: integer
          nullable: true
        weight:
          type: string
          nullable: true
      required: [vendorId, categoryId, name, price, stock]
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        unit:
          type: string
        stock:
          type: integer
        isBestSeller:
          type: boolean
        isSeasonal:
          type: boolean
        isActive:
          type: boolean
        tags:
          type: array
          items:
            type: string
        ingredients:
          type: string
          nullable: true
        expirationMonths:
          type: integer
          nullable: true
        weight:
          type: string
          nullable: true
...
    OrderItemDetail:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        productImage:
          type: string
          nullable: true
        price:
          type: number
        quantity:
          type: integer
        vendorId:
          type: string
          nullable: true
        vendorName:
          type: string
          nullable: true
      required: [id, productId, productName, price, quantity]
    OrderDetail:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        deliveryMethod:
          type: string
        deliveryAddress:
          type: string
          nullable: true
        deliveryFee:
          type: number
        subtotal:
          type: number
        total:
          type: number
        notes:
          type: string
          nullable: true
        referenceNumber:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemDetail"
        customerName:
          type: string
          nullable: true
        customerEmail:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, status, paymentMethod, paymentStatus, deliveryMethod, deliveryFee, subtotal, total, items, createdAt, updatedAt]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderDetail"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required: [orders, total, offset, limit]
    CreateOrderRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [gcash, maya, cod, bank_transfer, seven_eleven]
        deliveryMethod:
          type: string
          enum: [pickup, lalamove, jnt, lbc, hotel_dropoff]
        deliveryAddress:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        referenceNumber:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required: [paymentMethod, deliveryMethod, items]
    OrderItemInput:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
      required: [productId, quantity]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, out_for_delivery, delivered, cancelled]
        notes:
          type: string
          nullable: true
      required: [status]
...
    WishlistItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        createdAt:
          type: string
      required: [id, productId, product, createdAt]
    AddToWishlistRequest:
      type: object
      properties:
        productId:
          type: string
      required: [productId]
...